# fusecpt-backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API Documentation",
    "version": "1.0.0",
    "description": "API documentation for my Node.js + TypeScript backend"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/job": {
      "get": {
        "summary": "Get all jobs",
        "tags": ["Job"],
        "description": "Requires user to be logged in via cookie",
        "responses": {
          "200": {
            "description": "List of jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "jobs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/job/{id}": {
      "get": {
        "summary": "Get a job by ID",
        "tags": ["Job"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "patch": {
        "summary": "Update a job",
        "tags": ["Job"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Job ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a job",
        "tags": ["Job"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/api/job/create": {
      "post": {
        "summary": "Create a new job",
        "tags": ["Job"],
        "description": "Creates a new job.\nThis API accepts **multipart/form-data** because it includes an optional logo image upload.\nThe JSON job data must be sent inside a form field named **\"data\"** as a JSON string.\n",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["data"],
                "properties": {
                  "clientLogo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional client logo image file"
                  },
                  "data": {
                    "type": "string",
                    "description": "JSON object containing job fields.",
                    "example": "{\n  \"title\": \"Senior Developer\",\n  \"descriptionUrl\": \"https://example.com/job-description\",\n  \"clientName\": \"ABC Corp\",\n  \"clientTimezone\": \"PST\",\n  \"contractType\": \"Full-Time\",\n  \"overlapRequirement\": \"Partial\",\n  \"region\": [\"USA\", \"Canada\"],\n  \"minimumExperience\": \"3+ years\",\n  \"remarks\": \"Urgent role\",\n  \"status\": \"Open\"\n}\n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/job/history": {
      "get": {
        "summary": "Get global job activity history",
        "tags": ["Job"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "jobId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by specific job"
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by user who performed the action"
          }
        ],
        "responses": {
          "200": {
            "description": "Global job activity history"
          }
        }
      }
    },
...
  },
...
}